+++ src/SecsFrame/Transport/IHsmsTransportTimer.cs
@@ -1,0 +1,6 @@
+namespace SecsFrame;
+
+internal interface IHsmsTransportTimer : IDisposable
+{
+    void Change(TimeSpan dueTime);
+}