+++ src/SecsFrame/Transport/IHsmsTransportTimerFactory.cs
@@ -1,0 +1,6 @@
+namespace SecsFrame;
+
+internal interface IHsmsTransportTimerFactory
+{
+    IHsmsTransportTimer Create(Action callback);
+}